Update Attention-OCR
The Attention-OCR library that's linked to in this repo won't run on any of the recent Tensorflow versions or on ML engine. I maintain a fork that's being kept up-to-date and has some wrappers around the original model to provide better tooling. If you don't mind, it would make more sense to link to the fork instead. The original model is, of course, attributed to.